+++ packages/dropzone/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 15.0.0-rc.1
+
+### Patch Changes
+
+- Updated dependencies [b3669bf]
+  - react-magma-dom@5.2.0-rc.9
+
 ## 14.0.1-rc.0
 
 ### Patch Changes